Expose logsumexp as scipy.special.logsumexp.
scipy.misc.logsumexp is deprecated and appears slated to be removed entirely in scipy 1.3.
