+++ jax/scipy/special.py
@@ -16,10 +16,11 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import numpy as onp
 import scipy.special as osp_special
 
 from .. import lax
-from ..numpy.lax_numpy import _wraps, asarray
+from ..numpy.lax_numpy import _wraps, asarray, _reduction_dims, _constant_like
 
 
 # need to create new functions because _wraps sets the __name__ attribute
@@ -41,3 +42,17 @@ def expit(x):
   x = asarray(x)
   one = lax._const(x, 1)
   return lax.div(one, lax.add(one, lax.exp(lax.neg(x))))
+
+
+@_wraps(osp_special.logsumexp)
+def logsumexp(a, axis=None, b=None, keepdims=False, return_sign=False):
+  if b is not None or return_sign:
+    raise NotImplementedError("Only implemented for b=None, return_sign=False")
+  dims = _reduction_dims(a, axis)
+  shape = lax.subvals(onp.shape(a), zip(dims, (1,) * len(dims)))
+  dimadd = lambda x: lax.reshape(x, shape)
+  amax = lax.reduce(a, _constant_like(a, -onp.inf), lax.max, dims)
+  amax_singletons = dimadd(amax)
+  out = lax.add(lax.log(lax.reduce(lax.exp(lax.sub(a, amax_singletons)),
+                                   _constant_like(a, 0), lax.add, dims)), amax)
+  return dimadd(out) if keepdims else out
